requirements.txt
flask==3.1.0
flask_sqlalchemy==3.1.1
pymysql
python-dotenv
cryptography
requests